Renaming projects generic "CensusAPIService" to GeoNerves to match solution name
