+++ GeoNerves/CensusGeolocator.cs
@@ -1,4 +1,4 @@
-﻿using CensusAPIService.Models;
+﻿using GeoNerves.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,7 +7,7 @@ using System.Xml.Serialization;
 using Newtonsoft.Json;
 using System.Threading.Tasks;
 
-namespace CensusAPIService
+namespace GeoNerves
 {
     /// <summary>
     /// Address GeoLocator class
