+++ GeoNerves/Properties/AssemblyInfo.cs
